+++ web/app/containers/product-details/actions.js
@@ -59,29 +59,50 @@ export const goToWishlist = () => (dispatch, getState) => {
 const submitCartFormSelector = createPropsSelector({
     productId: getCurrentProductId,
     qty: getItemQuantity,
-    variations: getProductVariationCategories
+    variationCategories: getProductVariationCategories,
+    variants: getProductVariants
 })
 
+/**
+ * @param {Object} formValues form values of selected variant options
+ * @example
+ * {
+ *	"color": "8",
+ *	"size": "6"
+ * }
+ * @param {Array} variants all possible variants, array of {Objects}, each object has a values key.
+ * The values in this key will have the same shape / data as the above formValues if they match
+ * @return {Object} returns the variant object matching the selected form values
+ */
+const getProductFromFormValues = (formValues, variants) => {
+    return variants ? variants.find(({values}) => {
+        return Object.keys(formValues).every((key) => formValues[key] === values[key])
+    }) : null
+}
+
 export const submitCartForm = (formValues) => (dispatch, getStore) => {
-    const {productId, qty, variations} = submitCartFormSelector(getStore())
+    const {productId, qty, variationCategories, variants} = submitCartFormSelector(getStore())
     const path = window.location.pathname
     const itemIdMatch = path.match(/\/id\/(.*?)\/product_id\//)
 
-    if (variations) {
+    if (variationCategories) {
         const errors = {}
-        variations.forEach(({id, label}) => {
-            if (!formValues[id]) {
-                errors[id] = `Please select a ${label}`
+        variationCategories.forEach(({name, label}) => {
+            if (!formValues[name]) {
+                errors[name] = `Please select a ${label}`
             }
         })
         if (Object.keys(errors).length > 0) {
-            return Promise.reject(new SubmissionError(errors))
+            return Promise.reject(new SubmissionError({
+                _error: errors
+            }))
         }
     }
 
     dispatch(addToCartStarted())
+    const variant = getProductFromFormValues(formValues, variants)
 
-    return dispatch(itemIdMatch ? updateCartItem(itemIdMatch[1], qty, productId) : addToCart(productId, qty))
+    return dispatch(itemIdMatch ? updateCartItem(itemIdMatch[1], qty, productId, variant) : addToCart(productId, qty, variant))
         .then(() => {
             dispatch(setIsWishlistAdded(false))
             return dispatch(openModal(PRODUCT_DETAILS_ITEM_ADDED_MODAL, UI_NAME.addToCart))